+++ mountpoint-s3-fs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Add `tls_config` field on `s3::config::ClientConfig` so callers can configure a custom CA trust store through to the underlying S3 client. ([#1834](https://github.com/awslabs/mountpoint-s3/pull/1834))
+* Add `S3FilesystemConfig::allow_copy_rename` / `SuperblockConfig::allow_copy_rename` to emulate file rename on buckets without `RenameObject` via `CopyObject` + `DeleteObject` for objects up to 5 GiB, with session-local disable after unrecoverable authorization errors.
 * Allow mounting on top of `autofs` managed directories. ([#1762](https://github.com/awslabs/mountpoint-s3/pull/1762))
 * Fix memory limiter ignoring container cgroup memory limits, which could cause out-of-memory issues in memory-constrained containers. ([#1806](https://github.com/awslabs/mountpoint-s3/pull/1806))
 * Add `S3FilesystemConfig::content_type_detection` option to configure automatic content type inference for new uploads. When set to `ContentTypeDetection::Auto`, Mountpoint will infer the `Content-Type` of new objects based on their file extension. ([#1790](https://github.com/awslabs/mountpoint-s3/pull/1790))